Add FB Share Button
Background:
To set the preview image for a FB share, the image has to already be available
and set on a page in the og:image meta tag for FB to scrape. For this to work with all generated PNGs
we have to find a way to set this in the share URL so that the uniquely created infographic shows
in the preview. (Can discuss w/ Allen)

For now, the FB share button just uses the same header image for the city.

Changes:
-Adds FB Share button to share current page after generating the infographic
-Adds config option to turn off sending Push notifications
-Adds config option to set node executable path
-Adds config option for Facebook App Id which is required to use their JS SDK
-Allow setting some global constants in utils/constants.php.  Just setting main slash image for now.